Adding functionality to display command suggestions as you type.
The list of commands is parsed from the existing table, and a box with
documentation pops up under the text box as you type, showing the
possible commands, short forms, and operands and documentation.

Hitting Tab will also fill in the longest matching prefix string of all
the currently visible suggestions.

It's all in inline JavaScript with no external dependencies, for minimal
intrusiveness and minimal network traffic impact.